Remove mamba and expect, use pytest for coverage
