+++ .github/workflows/python-build.yml
@@ -31,6 +31,6 @@ jobs:
     - name: Run linters (flake8 and mypy)
       run: |
         make lint
-    - name: Run specs
+    - name: Run tests
       run: |
-        make lint spec test coverage
+        make lint coverage
