+++ pyproject.toml
@@ -8,17 +8,12 @@ authors = ["Attila Domokos <adomokos@gmail.com>"]
 python = "^3.6"
 
 [tool.poetry.dev-dependencies]
-mamba = "^0.11.1"
-expects = "^0.9.0"
 Mypy = "^0.782"
 flake8 = "^3.8.3"
 coverage = "^5.3"
 black = "^20.8b1"
 pytest = "^6.2.1"
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
+pytest-cov = "^2.10.1"
 
 [tool.pytest.ini_options]
 minversion = "6.0"
@@ -26,3 +21,7 @@ addopts = "-ra -q"
 testpaths = [
     "test",
 ]
+[build-system]
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
+
