--- spec/pyservice_spec.py
@@ -1,7 +1,0 @@
-from pyservice import __version__
-from mamba import description, it
-
-
-with description("Version") as self:
-    with it("verifies the package version"):
-        assert __version__ == "0.0.2"
